requirements.txt
numpy
statsmodels
pandas-profiling[notebook]
kaggle
